Update update-jira-tickets.yaml (#1229)
This PR updates the Update Jira Tickets GH Actions workflow to fetch all commits before determining JIRA tickets. 

The current behavior is only using the most recent commit on the branch because the checkout action does a shallow clone by default. Therefore, we have to unshallow the clone in order to get all the commit messages. I decided to do it this way, instead of doing a deep clone from the outset because there are several exit points for the job that would render the deep clone unnecessary. So I do it at the point where it's needed