M2-6457: Update the `/workspaces/{owner_id}/applets/{applet_id}/respondents/{respondent_id}` endpoint to account for multiple applets (#1264)
This PR is a simple change to account for multiple applets in the `/workspaces/{owner_id}/applets/{applet_id}/respondents/{respondent_id}` endpoint. The current implementation will return multiple results in an undefined order if the respondent being fetched has that role in multiple applets (and therefore has multiple entries in the `subjects` table).

By updating the join condition between `user_applet_accesses` and `subjects` to include both the `user_id` and `applet_id` columns, we ensure that the result set will have a single entry.