feat: Add/Update endpoint to retrieve activities assigned to a participant (M2-6223) (#1565)
This PR adds a new endpoint `/activities/applet/{applet_id}/subject/{subject_id}`, which allows the caller to fetch an object containing the list of activities and activity flows that have been assigned to the subject specified by `subject_id` in some way. Each activity or flow will have a non-nullable array property called `assignments` containing the list of assignments of that activity/flow where the given subject is the **respondent**.

It is yet undecided at a product level whether auto assigned activities/flows can be manually assigned, but the API currently allows this. Thus, it is possible for the data returned from this endpoint to contain an activity/flow with both `autoAssign` set to `true`, and a non-empty `assignments` property.

The caller of the endpoint must be authenticated as a user with one of the following user roles:
- Owner
- Manager
- Coordinator
- Reviewer (with access to the subject)