feat: Add Source subject Details to endpoints that return individual answer submissions (M2-8140) (#1645)
This PR updates the following endpoints to add a nullable `sourceSubject` property to their `answer` objects

- Get activity answer: `GET /answers/applet/{applet_id}/activities/{activity_id}/answers/{answer_id}`
- Get flow submission: `GET /answers/applet/{applet_id}/flows/{flow_id}/submissions/{submit_id}`

The structure of the `sourceSubject` is the same `SubjectReadResponse` as that returned from get subject endpoint (`GET /subjects/{subject_id}`). The only exception being that the `lastSeen` property will always be null in these cases.