feat: Update the workspace applet respondents list endpoint to include each respondent's role (M2-8465) (#1705)
This PR updates the get workspace respondents endpoints to add a `roles` property to each object in the `details` array. The `roles` property contains a list of roles the user has in a particular applet. This affects these endpoints:
- Workspace respondents list - `/workspaces/{owner_id}/respondents`
- Workspace applet respondents list -  `/workspaces/{owner_id}/applets/{applet_id}/respondents`

I've also added a `roles` property to each subject returned from these endpoints:
- Get subject - `/subjects/{subject_id}`
- Get target subjects by respondent - `/subjects/respondent/{respondent_subject_id}/activity-or-flow/{activity_or_flow_id}`