feat: Implement schedule history entities (M2-8417) (#1718)
This ticket makes several updates to support the implementation of schedule versioning and schedule history.

First off, the `events` table has been updated with the following columns:
- **version** - a string in the format `YYYYMMDD-n`, where `n` is the nth change on a given date
- **periodicity** - the `type` column from the `periodicity` table
- **start_date** - the `start_date` column from the `periodicity` table
- **end_date** - the `end_date` column from the `periodicity` table
- **selected_date** - the `selected_date` column from the `periodicity` table

Only the **version** column is required to support schedule versioning, but merging the periodicity columns with the events table simplifies a lot of the querying logic.

A number of new tables have also been created according to the ER diagram in the ticket:
- **event_histories** - keeps track of event versions
- **applet_events** - keeps track of which versions of a schedule are related to which versions of an applet
- **notification_histories** - The notifications tied to a particular schedule version
- **reminder_histories** - The reminder tied to a particular schedule version

As part of this implementation, I decided to forego the creation of the `user_event_histories`, `activity_event_histories`, and `flow_event_histories` tables and include the `user_id`, `activity_id`, and `activity_flow_id` columns respectively directly in the `event_histories` table. This merging simplifies queries, since these would have been one-to-one relationships with no extra data.

Finally, I updated the various event data structures being returned from the backend to include the version property. Since no data is being migrated as part of this ticket, the version property will always be `null` for now.