Fix the failed data migration
The data migration failed because of missing entries in the `applet_histories` table. The migration script assumed that there should be an entry in the applet_histories table for all the current applets but this proved not to be the case.

The script has been updated to restrict the migration to those applets that actually have entries in `applet_histories` using an explicit join condition. I've also done explicit joins on two subsequent migration statements