fix: Implement schedule change versioning + data migration (M2-8494) (#1733)
This PR is a re-issue of 9f0f19905d021f2f6b8876e81d68fd108a22139e to fix the failed data migration that resulted from the merge to develop of #1719.

The data migration failed because of missing entries in the applet_histories table. The migration script assumed that there should be an entry in the applet_histories table for all the current applets but this proved not to be the case. The script has been updated to restrict the migration to those applets that actually have entries in applet_histories using an explicit join condition.

I've also done explicit joins on two subsequent migration statements for notification_histories and reminder_histories, though I don't expect those to be needed.