fix: Fix linking of individual events to new applet versions in applet_events (M2-8743) (#1749)
This PR updates the query being used to find all the events linked to an applet when that applet is updated. This list of events is used as the base for which rows get inserted into `applet_events` whenever the applet version changes without changing any event versions.

Previously, the query being used was excluding individual events by adding the condition `WHERE user_id = NULL`