chore: Restore *_event tables (M2-8495) (#1752)
This PR restores the following tables that were previous removed as part of #1723:
- activity_events
- flow_events
- user_events

This restoration is optional and should only run if the tables are missing. So it should run on dev, but do nothing on UAT.

The rest of the code remains unchanged, and this change is made on Jody's suggestion to split this across two releases. There will be a follow-up PR that drops these tables in about a week or so