fix: Update event_history_id validation (M2-8482) (M2-8838) (#1777)
This PR updates the validation of the `event_history_id` property in the answer data to work for both activities and flows. Previously it was only working for activities.

The reason it failed is because of an assumption that the `event` data contains both the activity and the flow ID, but this is not the case, it only has one or the other. Thus, we can only perform one comparison based on what type of submission is being done.