feat: Create Applet Seed CLI (M2-8991) (#1802)
This PR creates a V1 CLI command to allow a developer or QA tested to populate some seed data in the database. The
motivation for this command is to create an easy to read DSL that allows QA testers to create applet data in the past
for testing purposes. It is a V1 implementation, and does not yet support several structures, such as activity flows. In
the future, the format may be improved with new versions, while maintaining backwards compatibility.

This PR also creates a couple database migrations that need to be run before the CLI command will work. I found it
necessary to add update cascades to several foreign key columns to allow updating the primary key of their referenced
fields. This allows us to specify the UUID of an applet, activity, etc. in the seed file and have it be reflected in the
database.