Updates dependencies
Updates `taskiq-redis`, `allure-pytest`, `pytest-asyncio`, `pytest-cov`, and `types-pytz`
dependencies in `Pipfile` and `Pipfile.lock`.
Also, updates the hashes and markers in `Pipfile.lock` to reflect the new versions.

Updates dependencies in Pipfile

Pin several dependencies to newer versions in Pipfile
and Pipfile.lock to pick up security fixes and new features.

Updates dependencies in Pipfile.lock

This commit updates the dependencies in the Pipfile.lock file
to their latest versions, including:

- boto3
- botocore
- bytecode
- ddtrace
- gitignore-parser
- google-api-python-client
- google-auth
- google-cloud-firestore
- googleapis-common-protos
- greenlet
- opentelemetry-api
- rsa
- sentry-sdk
- starlette
- tomli
- uvloop
- yarl
- prompt-toolkit

This ensures that the project is using the most recent and
secure versions of its dependencies.