feat: Remove date filtering from schedule history endpoints (M2-8485) (#1815)
This PR removes the date query parameters from the schedule history endpoints because they are not needed on the backend. The way these data parameters are intended to be used is to exclude rows from the exported CSV file. These rows are enumerated such that one event object in the API response may produce 0 or more rows in the export file.

However, these rows are only generated on the frontend. When the API uses dates to filter event objects from the response (based on creation time), it could be removing valid events that would have otherwise generated rows that fall within the date range.

This change is required to properly evaluate bugs being spawned from M2-8485