chore: Revise applet-event version linking logic (M2-8973) (M2-8983) (M2-9047) (#1820)
This PR revises the logic that links versions of schedule events and applets together in the `applet_events` table. The current logic has a bug where an update to an applet that deletes an activity/flow results in the events for that activity/flow being linked to the new version of the applet in `applet_events`.

The root cause is that the linking action was being done before the events for the activity/flow were marked as deleted. This PR moves the linking action to the end of the update method to account for this.

I also took the opportunity to fix the `subject_id` field filter in the schedule history endpoints. The `subject_id` filter was missing the None condition because similar to `user_id`, when filtering by `subject_id` the event object should be returned when the field matches, or when it is null