chore: EMA Exports Extras (M2-8958) (M2-9087) (M2-9088) (#1827)
This PR makes a few changes to the schedule history, flow item history, and respondent endpoints to enable certain features of the EMA data files exports in the admin panel.

Schedule History

The `activity_or_flow_ids` query parameter has been added to the schedule history endpoint to allow for filtering the results by a list of activity and flow IDs. When specified, the events returned should only be for those activities and flows specified in this list.

Flow Item History

The `flow_ids` query parameter has been added to the flow item history endpoint to allow filtering the results by a list of flow IDs. When specified, the history returned should only include a subset of activity flows in this list.

Workspace Respondents

An `acceptedAt` property has been added to the `invitation` property on each applet detail object. This value should represent the date when an invitation was approved (if it is in approved status), and otherwise be `null`

This new property supports the exclusion of respondents from the schedule history file for dates before they were added to the applet