Refactors Oneup Health integration
Updates the Oneup Health integration to include the activity ID when creating and retrieving users.
This ensures uniqueness when the same submit ID is used in multiple activities.
Also, it includes the app_user_id in the OneupHealthToken.
A unique short ID is generated using a hash of the submit ID and activity ID.