fix: EMA Fast Follows 2 (M2-9108) (M2-9161) (#1848)
This PR updates a couple API endpoints in preparation for a corresponding PR that resolves the captioned tickets.

#### M2-9108: The whole schedule_history associated with the removed participant is also removed

This ticket is about retaining schedule history in the export file for participants whose subjects are soft deleted. The FE uses the workspace respondents endpoint to fetch the complete list of participants in the applet, and generate an enumerated export file for each of them for the selected date range. This PR updates the workspace respondents endpoint to optionally include soft deleted participants, and adds two key properties to the response:
- `subjectIsDeleted` - Whether the subject is soft deleted within a given applet
- `subjectUpdatedAt` - The date when the last change was made to this subject (approximating the deletion date)

This will allow the FE to generate an enumerated schedule history for soft deleted participants up to the day when they were removed from the applet.

> [!NOTE]
> This only applies to soft deletion, which is when the admin removes the participant without removing their data. When the admin chooses to remove the data, the subject is hard-deleted and will be removed from the workspace respondents api response

#### M2-9161: 500 error appears in the Console after starting export in Migrated applet

This is the result of migrated applets having `NULL` values in the `start_date` and `start_time` database columns for always available schedule events. Current applets always populate these fields for all periodicities. I have accounted for this in the schedule history endpoint by defaulting the values of these fields to those used by the BE when creating events with an always available periodicity: namely `'00:00:00'` and `'23:59:00'` respectively.