fix: Clean up subjects table is_deleted column (M2-9161) (M2-9250) (#1855)
This PR adds a data migration that changes `NULL` values in the `is_deleted` column of the `subjects` table to `FALSE`. This is in line with how the soft_exists method is implemented, where it returns records where `is_deleted IS NOT TRUE`.

I've added an extra column called `is_deleted_null` to track these changes, in case we need to go back