Enables EHR data export with filenames
Adds the ability to include EHR data filenames in answer exports.
This enhancement provides users with a more complete dataset,
facilitating easier identification and access to associated EHR
files. A new filter is introduced, and a service method is added
to retrieve the EHR filenames based on activity and submit IDs.