feat: Rename MindLogger references to Curious (M2-9140) (#1862)
* feat: Rename MindLogger references to Curious (M2-9140)

Updates all  end-user facing references of "MindLogger" to "Curious" in the codebase,
reflecting the application's rebranding. This includes changes to
branding assets, email templates, documentation, and default
configuration values.

* Updates email does not exist message

Corrects a message to reflect the Curious branding.

The previous message incorrectly referred to "MindLogger"
instead of the current "Curious" branding.

* Removes unused email address error

Removes the `EmailAddressError` exception class as it is no longer being used.