feat: Enhances OneUp Health integration for document retrieval (M2-8883) (#1879)
* Enhances OneUp Health integration for document retrieval

Improves the OneUp Health integration by adding functionality
to download and store patient documents associated with EHR data.

This change introduces the ability to fetch documents referenced
in DocumentReference resources, store them in a zip file,
and upload the zip to the EHR storage.

It also introduces generic file uploading and listing capabilities
in the EHR storage, along with helper methods.

* Adds metadata to EHR answers

Adds metadata to the EHR answers table to store information about the uploaded zip files, including their names and sizes.

This allows for better tracking and management of the EHR data stored in the system.

* Adds assertion to task ingest user data

Adds an assertion to check that the result of the
`ingest_user_data` function is not None. This ensures that
the task completes successfully and returns a valid result.

* Update src/apps/integrations/oneup_health/service/oneup_health.py

Co-authored-by: Farmer Paul <paul.hh@metalab.com>

* Refactors EHR metadata handling

Updates EHR metadata to use a dedicated data model.

This change introduces `EHRFileMetadata` and `EHRFileTypeEnum`
to provide more structured and type-safe handling of EHR file
metadata, replacing the previous use of dictionaries. This improves
code clarity and maintainability.

* Adds provider name to EHR document filename

Ensures that the EHR document filename includes the
healthcare provider's name (or ID if name unavailable)
to improve identification and organization.

* Adds buckets for answers and operations

Extends MinIO bucket creation to include buckets for answers and operations, in addition to the existing media bucket.
This allows for a more organized storage structure within MinIO and improves separation of concerns.

* Adds a TODO item to optimize zip file creation.

Adds a TODO item to address potential memory issues
when creating zip files, especially when dealing with
large documents or numerous files.

* Refactors EHR file upload process

Streamlines the EHR file upload process by moving the base path
generation logic into the `EHRStorage.upload_file` method.
This change improves code maintainability and reduces redundancy.

---------

Co-authored-by: Farmer Paul <paul.hh@metalab.com>